[ADD] d_c_website_sale: Reminder for abandoned carts
Add functionality to send email to the owners of abandoned carts before
the end of a time frame.

In the website configuration, this adds  the number of hours before the
end of a time frame that have to be taken into account to send the
reminder email.

This also adds a cron that check if there are abandoned carts each days.