[FIX] add call to product_id_change function after sol creation.
In order to avoid problem relating to translation we add the call to the
product_id_change function to take into account all the context tied to
the sale order. It has also the advantage to take the price list into
account which wasn't the case in the previous code.